+++ database/database.py
@@ -1,46 +1,34 @@
 import mysql.connector as mariadb
-from enum import Enum
 import os
-
-class Role(Enum):
-    USER = 0
-    TEACHER = 1
-    ADMIN = 2
-
-class AccountStatus(Enum):
-    UNVERIFIED = 0
-    VERIFIED = 1
-    REPORTED = 2
-    BANNED = 3
+from database import accountStatus, role
 
 class Database():
 
-    #def __init__(self, host, port, username, password, database):
-    #    self.host, self.port, self.username, self.password, self.database = host, port, username, password, database
-    #    self.connection, self.cursor = None, None
-
-    def __int__(self):
-        self.host, self.port, self.username, self.password, self.database = "localhost", "3306", os.environ["MYSQL_USER"], os.environ["MYSQL_PASSWORD"], os.environ["MYSQL_RANDOM_ROOT_PASSWORD"]
+    def __init__(self):
+        self.host, self.port, self.username, self.password, self.database = os.environ["DATABASE_HOSTNAME"], str(os.environ["DATABASE_PORT"]), os.environ["DATABASE_USERNAME"], os.environ["DATABASE_PASSWORD"], os.environ["DATABASE_DATABASE"]
         self.connection, self.cursor = None, None
 
     def connect(self):
-        self.connection = mariadb.connect(host=self.host, port=str(self.port), user=self.username, password=self.password, database=self.database)
-        self.cursor = self.connection.cursor(prepared=True)
-        self.connection.commit()
-        print("Database connected to '" + self.host + ":" + str(self.port) + "'")
+        try:
+            self.connection = mariadb.connect(host=self.host, port=str(self.port), user=self.username, password=self.password, database=self.database)
+            self.cursor = self.connection.cursor(prepared=True)
+            self.connection.commit()
+            print("[DATABASE] Database connected to '" + self.host + ":" + str(self.port) + "'")
+        except: print("[DATABASE] Error: Failed to connect!")
 
     def isConnected(self):
         return self.connection is not None and self.connection.is_connected()
 
     def checkConnection(self):
-        if not self.isConnected():
+        try:
+            self.connection.ping(reconnect=True, attempts=3, delay=1)
+        except mariadb.Error:
             self.connect()
 
     def execute(self, sql, values=()):
-        try:
-            return self.cursor.execute(sql, values)
+        try: return self.cursor.execute(sql, values)
         except mariadb.Error as error:
-            print("Error:" + str(error))
+            print("[DATABASE] Error:" + str(error))
 
     def existUser(self, email):
         self.checkConnection()
@@ -48,7 +36,7 @@ class Database():
         self.execute('SELECT userId FROM Users WHERE email = ?', (email,))
         return self.cursor.fetchone() is not None
 
-    def createUser(self, firstName, lastName, email, hashedPassword, role: Role):
+    def createUser(self, firstName, lastName, email, hashedPassword, role: role.Role):
         self.checkConnection()
         if not self.isConnected(): return None
         self.execute('INSERT INTO Users (firstName, lastName, email, password, accountStatus, role) VALUES (?, ?, ?, ?, ?, ?)', (firstName, lastName, email, hashedPassword, 0, role.value))
@@ -76,11 +64,11 @@ class Database():
         for row in result:
             if type(row) == bytearray: info.append(row.decode("utf-8"))
             else: info.append(row)
-        info[5] = AccountStatus(info[5])
-        info[7] = Role(info[7])
+        info[5] = accountStatus.AccountStatus(info[5])
+        info[7] = role.Role(info[7])
         return info
 
-    def updateUserInfo(self, userId, firstName=None, lastName=None, email=None, hashedPassword=None, accountStatus=None):
+    def updateUserInfo(self, userId, firstName=None, lastName=None, email=None, hashedPassword=None, accountStatus: accountStatus.AccountStatus=None):
         self.checkConnection()
         if not self.isConnected(): return None
         if firstName is not None:
@@ -92,7 +80,7 @@ class Database():
         if hashedPassword is not None:
             self.execute('UPDATE Users SET password = ? WHERE userId = ?', (hashedPassword, userId))
         if accountStatus is not None:
-            self.execute('UPDATE Users SET accountStatus = ? WHERE userId = ?', (accountStatus, userId))
+            self.execute('UPDATE Users SET accountStatus = ? WHERE userId = ?', (accountStatus.value, userId))
         self.connection.commit()
 
     def getUserIdByEmail(self, email):
