+++ database/role.py
@@ -1,0 +1,6 @@
+from enum import Enum
+
+class Role(Enum):
+    USER = 0
+    TEACHER = 1
+    ADMIN = 2
